Add get registry function for metrics registries, and use them.
